# lib/api-zod/src/generated/types/decisao.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Decisao {
  id: number;
  data: Date;
  titulo: string;
  valor: number;
  categoria: string;
}